--- packages-private/common-utils/source/valibot/schemas/functions.ts
@@ -1,465 +1,0 @@
-/*
- * Copyright 2026 Adobe. All rights reserved.
- * This file is licensed to you under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License. You may obtain a copy
- * of the License at http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under
- * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
- * OF ANY KIND, either express or implied. See the License for the specific language
- * governing permissions and limitations under the License.
- */
-
-import * as v from "valibot";
-
-import { returnsSyncOrAsync } from "../actions/returns-sync-or-async";
-import { withPrefixedMessage } from "../messages";
-
-const NO_ARGS_MESSAGE = "Expected no arguments for this function";
-const INVALID_ARGS_MESSAGE =
-  "The given arguments do not match the expected function signature";
-
-const NO_OUTPUT_MESSAGE = "Expected no output from this function";
-const INVALID_OUTPUT_MESSAGE =
-  "The function does not return the expected output";
-
-const INVALID_FUNCTION_MESSAGE = "The value is not a valid function";
-
-function createArgumentIssuePrefix(issue: v.BaseIssue<unknown>) {
-  const dotPath = v.getDotPath(issue);
-
-  if (!dotPath) {
-    return `${INVALID_ARGS_MESSAGE} → arguments`;
-  }
-
-  // For args, Valibot doesn't know the argument names, so the path is based on the argument index
-  // (e.g. "0" for the first argument) and any nested properties (e.g. "0.name" for a "name" property on the first argument).
-  // This function converts that into a more user-friendly message like "argument 1" instead of "arguments.0".
-  const [firstSegment, ...restSegments] = dotPath.split(".");
-  const argumentIndex = Number(firstSegment);
-
-  if (!Number.isInteger(argumentIndex)) {
-    return `${INVALID_ARGS_MESSAGE} → arguments.${dotPath}`;
-  }
-
-  const nestedPath =
-    restSegments.length > 0 ? `.${restSegments.join(".")}` : "";
-
-  return `${INVALID_ARGS_MESSAGE} → arguments[${argumentIndex}]${nestedPath}`;
-}
-
-/** A schema for a function that may return either a direct value or a promise-like value. */
-export type AnyTupleSchema = v.TupleSchema<
-  v.TupleItems,
-  v.ErrorMessage<v.TupleIssue> | undefined
->;
-
-/**
- * Options for creating a function schema, including optional argument and output schemas and a custom message.
- * @template ArgsSchema The schema for the function arguments, which must be a tuple schema or undefined.
- * @template OutputSchema The schema for the function output, which must be a Valibot generic schema or undefined.
- */
-export type FunctionSchemaOptions<
-  ArgsSchema extends AnyTupleSchema | undefined,
-  OutputSchema extends v.GenericSchema | undefined,
-> = {
-  args?: ArgsSchema;
-  output?: OutputSchema;
-  message?: string;
-};
-
-/**
- * Convenience type for an options object which only allows configuring the args for a function schema.
- * @template ArgsSchema The schema for the function arguments, which must be a tuple schema.
- */
-export type OnlyArgsFunctionSchemaOptions<
-  ArgsSchema extends AnyTupleSchema | undefined,
-> = Omit<FunctionSchemaOptions<ArgsSchema, undefined>, "output">;
-
-/** A schema for a simple synchronous function (no args, no output) */
-function createSyncFunctionSchema(message = INVALID_FUNCTION_MESSAGE) {
-  return v.pipe(
-    v.function(message),
-    v.args(withPrefixedMessage(v.tuple([]), NO_ARGS_MESSAGE)),
-    v.returns(withPrefixedMessage(v.void(), NO_OUTPUT_MESSAGE)),
-  );
-}
-
-/**
- * A schema for a synchronous function with arguments.
- * @param args The schema for the function arguments.
- */
-function createSyncFunctionSchemaWithArgs<ArgsSchema extends AnyTupleSchema>(
-  args: ArgsSchema,
-  message = INVALID_FUNCTION_MESSAGE,
-) {
-  return v.pipe(
-    v.function(message),
-    v.args(withPrefixedMessage(args, createArgumentIssuePrefix)),
-    v.returns(withPrefixedMessage(v.void(), NO_OUTPUT_MESSAGE)),
-  );
-}
-
-/**
- * A schema for a synchronous function with output.
- * @param output The schema for the function output.
- */
-function createSyncFunctionSchemaWithOutput<
-  OutputSchema extends v.GenericSchema,
->(output: OutputSchema, message = INVALID_FUNCTION_MESSAGE) {
-  return v.pipe(
-    v.function(message),
-    v.args(withPrefixedMessage(v.tuple([]), NO_ARGS_MESSAGE)),
-    v.returns(withPrefixedMessage(output, INVALID_OUTPUT_MESSAGE)),
-  );
-}
-
-/**
- * A schema for a synchronous function with both arguments and output.
- * @param args A schema for the function arguments.
- * @param output A schema for the function output.
- */
-function createSyncFunctionSchemaWithArgsAndOutput<
-  ArgsSchema extends AnyTupleSchema,
-  OutputSchema extends v.GenericSchema,
->(args: ArgsSchema, output: OutputSchema, message = INVALID_FUNCTION_MESSAGE) {
-  return v.pipe(
-    v.function(message),
-    v.args(withPrefixedMessage(args, createArgumentIssuePrefix)),
-    v.returns(withPrefixedMessage(output, INVALID_OUTPUT_MESSAGE)),
-  );
-}
-
-// Overload for function schema with both arguments and output
-export function syncFunctionSchema<
-  const ArgsSchema extends AnyTupleSchema,
-  const OutputSchema extends v.GenericSchema,
->(options: {
-  args: ArgsSchema;
-  output: OutputSchema;
-  message?: string;
-}): ReturnType<
-  typeof createSyncFunctionSchemaWithArgsAndOutput<ArgsSchema, OutputSchema>
->;
-
-// Overload for sync function schema with only arguments
-export function syncFunctionSchema<
-  const ArgsSchema extends AnyTupleSchema,
->(options: {
-  args: ArgsSchema;
-  output?: undefined;
-  message?: string;
-}): ReturnType<typeof createSyncFunctionSchemaWithArgs<ArgsSchema>>;
-
-// Overload for sync function schema with only output
-export function syncFunctionSchema<
-  const OutputSchema extends v.GenericSchema,
->(options: {
-  args?: undefined;
-  output: OutputSchema;
-  message?: string;
-}): ReturnType<typeof createSyncFunctionSchemaWithOutput<OutputSchema>>;
-
-// Overload for sync function schema with no arguments and no output
-export function syncFunctionSchema(options?: {
-  args?: undefined;
-  output?: undefined;
-  message?: string;
-}): ReturnType<typeof createSyncFunctionSchema>;
-
-/**
- * Creates a Valibot schema for a synchronous function with optional argument and output schemas.
- * @param options An object describing the arguments and output of the function schema to create.
- *
- * @example
- * ```ts
- * const schema = syncFunctionSchema({
- *   args: v.tuple([v.string()]),
- *   output: v.number(),
- * });
- * ```
- */
-export function syncFunctionSchema<
-  const ArgsSchema extends AnyTupleSchema | undefined,
-  const OutputSchema extends v.GenericSchema | undefined,
->(options?: FunctionSchemaOptions<ArgsSchema, OutputSchema>) {
-  if (!options) {
-    return createSyncFunctionSchema();
-  }
-
-  if (options.args && options.output) {
-    return createSyncFunctionSchemaWithArgsAndOutput(
-      options.args,
-      options.output,
-      options.message,
-    );
-  }
-
-  if (options.args) {
-    return createSyncFunctionSchemaWithArgs(options.args, options.message);
-  }
-
-  if (options.output) {
-    return createSyncFunctionSchemaWithOutput(options.output, options.message);
-  }
-
-  return createSyncFunctionSchema(options?.message);
-}
-
-/** A schema for an asynchronous function with optional argument and output schemas. */
-function createAsyncFunctionSchema(message = INVALID_FUNCTION_MESSAGE) {
-  return v.pipe(
-    v.function(message),
-    v.args(withPrefixedMessage(v.tuple([]), NO_ARGS_MESSAGE)),
-    v.returnsAsync(withPrefixedMessage(v.void(), NO_OUTPUT_MESSAGE)),
-  );
-}
-
-/** A schema for an asynchronous function with arguments. */
-function createAsyncFunctionSchemaWithArgs<ArgsSchema extends AnyTupleSchema>(
-  args: ArgsSchema,
-  message = INVALID_FUNCTION_MESSAGE,
-) {
-  return v.pipe(
-    v.function(message),
-    v.args(withPrefixedMessage(args, createArgumentIssuePrefix)),
-    v.returnsAsync(withPrefixedMessage(v.void(), NO_OUTPUT_MESSAGE)),
-  );
-}
-
-/** A schema for an asynchronous function with output. */
-function createAsyncFunctionSchemaWithOutput<
-  OutputSchema extends v.GenericSchema,
->(output: OutputSchema, message = INVALID_FUNCTION_MESSAGE) {
-  return v.pipe(
-    v.function(message),
-    v.args(withPrefixedMessage(v.tuple([]), NO_ARGS_MESSAGE)),
-    v.returnsAsync(withPrefixedMessage(output, INVALID_OUTPUT_MESSAGE)),
-  );
-}
-
-/** A schema for an asynchronous function with both arguments and output. */
-function createAsyncFunctionSchemaWithArgsAndOutput<
-  ArgsSchema extends AnyTupleSchema,
-  OutputSchema extends v.GenericSchema,
->(args: ArgsSchema, output: OutputSchema, message = INVALID_FUNCTION_MESSAGE) {
-  return v.pipe(
-    v.function(message),
-    v.args(withPrefixedMessage(args, createArgumentIssuePrefix)),
-    v.returnsAsync(withPrefixedMessage(output, INVALID_OUTPUT_MESSAGE)),
-  );
-}
-
-// Overload for async function schema with both arguments and output
-export function asyncFunctionSchema<
-  const ArgsSchema extends AnyTupleSchema,
-  const OutputSchema extends v.GenericSchema,
->(options: {
-  args: ArgsSchema;
-  output: OutputSchema;
-  message?: string;
-}): ReturnType<
-  typeof createAsyncFunctionSchemaWithArgsAndOutput<ArgsSchema, OutputSchema>
->;
-
-// Overload for async function schema with only arguments
-export function asyncFunctionSchema<
-  const ArgsSchema extends AnyTupleSchema,
->(options: {
-  args: ArgsSchema;
-  output?: undefined;
-  message?: string;
-}): ReturnType<typeof createAsyncFunctionSchemaWithArgs<ArgsSchema>>;
-
-// Overload for async function schema with only output
-export function asyncFunctionSchema<
-  const OutputSchema extends v.GenericSchema,
->(options: {
-  args?: undefined;
-  output: OutputSchema;
-  message?: string;
-}): ReturnType<typeof createAsyncFunctionSchemaWithOutput<OutputSchema>>;
-
-// Overload for async function schema with no arguments and no output
-export function asyncFunctionSchema(options?: {
-  args?: undefined;
-  output?: undefined;
-  message?: string;
-}): ReturnType<typeof createAsyncFunctionSchema>;
-
-/**
- * Creates a Valibot schema for an asynchronous function with optional argument and output schemas.
- * @param options An object describing the arguments and output of the function schema to create.
- *
- * @example
- * ```ts
- * const schema = asyncFunctionSchema({
- *   args: v.tuple([v.string()]),
- *   output: v.number(),
- * });
- * ```
- */
-export function asyncFunctionSchema<
-  const ArgsSchema extends AnyTupleSchema | undefined,
-  const OutputSchema extends v.GenericSchema | undefined,
->(options?: FunctionSchemaOptions<ArgsSchema, OutputSchema>) {
-  if (!options) {
-    return createAsyncFunctionSchema();
-  }
-
-  if (options.args && options.output) {
-    return createAsyncFunctionSchemaWithArgsAndOutput(
-      options.args,
-      options.output,
-      options.message,
-    );
-  }
-
-  if (options.args) {
-    return createAsyncFunctionSchemaWithArgs(options.args, options.message);
-  }
-
-  if (options.output) {
-    return createAsyncFunctionSchemaWithOutput(options.output, options.message);
-  }
-
-  return createAsyncFunctionSchema(options?.message);
-}
-
-/** A schema for a function that can be either synchronous or asynchronous (no args, no output) */
-function createSyncOrAsyncFunctionSchema(message = INVALID_FUNCTION_MESSAGE) {
-  return v.pipe(
-    v.function(message),
-    v.args(withPrefixedMessage(v.tuple([]), NO_ARGS_MESSAGE)),
-    returnsSyncOrAsync(withPrefixedMessage(v.void(), NO_OUTPUT_MESSAGE)),
-  );
-}
-
-/**
- * A schema for a function that can be either synchronous or asynchronous with arguments.
- * @param args The schema for the function arguments.
- */
-function createSyncOrAsyncFunctionSchemaWithArgs<
-  ArgsSchema extends AnyTupleSchema,
->(args: ArgsSchema, message = INVALID_FUNCTION_MESSAGE) {
-  return v.pipe(
-    v.function(message),
-    v.args(withPrefixedMessage(args, createArgumentIssuePrefix)),
-    returnsSyncOrAsync(withPrefixedMessage(v.void(), NO_OUTPUT_MESSAGE)),
-  );
-}
-
-/**
- * A schema for a function that can be either synchronous or asynchronous with output.
- * @param output The schema for the function output.
- */
-function createSyncOrAsyncFunctionSchemaWithOutput<
-  OutputSchema extends v.GenericSchema,
->(output: OutputSchema, message = INVALID_FUNCTION_MESSAGE) {
-  return v.pipe(
-    v.function(message),
-    v.args(withPrefixedMessage(v.tuple([]), NO_ARGS_MESSAGE)),
-    returnsSyncOrAsync(withPrefixedMessage(output, INVALID_OUTPUT_MESSAGE)),
-  );
-}
-
-/**
- * A schema for a function that can be either synchronous or asynchronous with both arguments and output.
- * @param args The schema for the function arguments.
- * @param output The schema for the function output.
- */
-function createSyncOrAsyncFunctionSchemaWithArgsAndOutput<
-  ArgsSchema extends AnyTupleSchema,
-  OutputSchema extends v.GenericSchema,
->(args: ArgsSchema, output: OutputSchema, message = INVALID_FUNCTION_MESSAGE) {
-  return v.pipe(
-    v.function(message),
-    v.args(withPrefixedMessage(args, createArgumentIssuePrefix)),
-    returnsSyncOrAsync(withPrefixedMessage(output, INVALID_OUTPUT_MESSAGE)),
-  );
-}
-
-// Overload for sync or async function schema with both arguments and output
-export function syncOrAsyncFunctionSchema<
-  const ArgsSchema extends AnyTupleSchema,
-  const OutputSchema extends v.GenericSchema,
->(options: {
-  args: ArgsSchema;
-  output: OutputSchema;
-  message?: string;
-}): ReturnType<
-  typeof createSyncOrAsyncFunctionSchemaWithArgsAndOutput<
-    ArgsSchema,
-    OutputSchema
-  >
->;
-
-// Overload for sync or async function schema with only arguments
-export function syncOrAsyncFunctionSchema<
-  const ArgsSchema extends AnyTupleSchema,
->(options: {
-  args: ArgsSchema;
-  output?: undefined;
-  message?: string;
-}): ReturnType<typeof createSyncOrAsyncFunctionSchemaWithArgs<ArgsSchema>>;
-
-// Overload for sync or async function schema with only output
-export function syncOrAsyncFunctionSchema<
-  const OutputSchema extends v.GenericSchema,
->(options: {
-  args?: undefined;
-  output: OutputSchema;
-  message?: string;
-}): ReturnType<typeof createSyncOrAsyncFunctionSchemaWithOutput<OutputSchema>>;
-
-// Overload for sync or async function schema with no arguments and no output
-export function syncOrAsyncFunctionSchema(options?: {
-  args?: undefined;
-  output?: undefined;
-  message?: string;
-}): ReturnType<typeof createSyncOrAsyncFunctionSchema>;
-
-/**
- * Creates a Valibot schema for a function that may return either a direct value or a promise-like value.
- * @param options An object describing the arguments and output of the function schema to create.
- *
- * @example
- * ```ts
- * const schema = syncOrAsyncFunctionSchema({
- *   args: v.tuple([v.string()]),
- *   output: v.number(),
- * });
- * ```
- */
-export function syncOrAsyncFunctionSchema<
-  const ArgsSchema extends AnyTupleSchema | undefined,
-  const OutputSchema extends v.GenericSchema | undefined,
->(options?: FunctionSchemaOptions<ArgsSchema, OutputSchema>) {
-  if (!options) {
-    return createSyncOrAsyncFunctionSchema();
-  }
-
-  if (options.args && options.output) {
-    return createSyncOrAsyncFunctionSchemaWithArgsAndOutput(
-      options.args,
-      options.output,
-      options.message,
-    );
-  }
-
-  if (options.args) {
-    return createSyncOrAsyncFunctionSchemaWithArgs(
-      options.args,
-      options.message,
-    );
-  }
-
-  if (options.output) {
-    return createSyncOrAsyncFunctionSchemaWithOutput(
-      options.output,
-      options.message,
-    );
-  }
-
-  return createSyncOrAsyncFunctionSchema(options?.message);
-}
